changed data type from integer to float for unit-of-measurement

+++ db/migrate/20210608200233_create_standardized_measurements.rb
@@ -3,7 +3,7 @@ class CreateStandardizedMeasurements < ActiveRecord::Migration[6.1]
     create_table :standardized_measurements do |t|
       t.string :name
       t.string :type
-      t.integer :unit_of_measurement
+      t.float :unit_of_measurement
 
       t.timestamps
     end
