+++ db/migrate/20210608200426_create_strange_measurements.rb
@@ -3,7 +3,7 @@ class CreateStrangeMeasurements < ActiveRecord::Migration[6.1]
     create_table :strange_measurements do |t|
       t.string :name
       t.string :type
-      t.integer :unit_of_measurement
+      t.float :unit_of_measurement
       t.string :source
 
       t.timestamps
